Fixed scores, rolling back failed Google assessment, and other minor issues.

+++ src/main/java/ca/canada/digital/search/assessment/service/TermAssessmentService.java
@@ -37,13 +37,13 @@ public class TermAssessmentService {
                 .filter(term -> term.getSearchType() == TermAssessment.SearchType.GOOGLE)
                 .count();
 
-        long internalPasses = termAssessments.stream()
+        float internalPasses = termAssessments.stream()
                 .filter(term -> term.getSearchType() == TermAssessment.SearchType.INTERNAL && term.getPass())
                 .count();
-        long internalSpecificPasses = termAssessments.stream()
+        float internalSpecificPasses = termAssessments.stream()
                 .filter(term -> term.getSearchType() == TermAssessment.SearchType.INTERNAL_SPECIFIC && term.getPass())
                 .count();
-        long googlePasses = termAssessments.stream()
+        float googlePasses = termAssessments.stream()
                 .filter(term -> term.getSearchType() == TermAssessment.SearchType.GOOGLE && term.getPass())
                 .count();
 
@@ -61,7 +61,7 @@ public class TermAssessmentService {
         UrlAssessmentResponse urlAssessmentResponse = new UrlAssessmentResponse();
         urlAssessmentResponse.setUrl(url);
         if (internalCount > 0) {
-            urlAssessmentResponse.setInternalScore(internalPasses > 0 ? String.format("%d%%", (internalPasses / internalCount) * 100) : "0%");
+            urlAssessmentResponse.setInternalScore(internalPasses > 0 ? String.format("%.2f%%", (float) (internalPasses / internalCount) * 100) : "0%");
             urlAssessmentResponse.setInternalTerms(internalTerms);
 
             LanguageProcess lp;
@@ -73,7 +73,7 @@ public class TermAssessmentService {
             urlAssessmentResponse.setHighlightedMetadata(metadataHighlights);
         }
         if (internalSpecificCount > 0) {
-            urlAssessmentResponse.setInternalSpecificScore(internalSpecificPasses > 0 ? String.format("%d%%", (internalSpecificPasses / internalSpecificCount) * 100) : "0%");
+            urlAssessmentResponse.setInternalSpecificScore(internalSpecificPasses > 0 ? String.format("%.2f%%", (float) (internalSpecificPasses / internalSpecificCount) * 100) : "0%");
             urlAssessmentResponse.setInternalSpecificTerms(termAssessments.stream()
                     .filter(term -> term.getSearchType() == TermAssessment.SearchType.INTERNAL_SPECIFIC)
                     .sorted(
@@ -85,7 +85,7 @@ public class TermAssessmentService {
                     .collect(Collectors.toList()));
         }
         if (googleCount > 0) {
-            urlAssessmentResponse.setGoogleScore(googlePasses > 0 ? String.format("%d%%", (googlePasses / googleCount) * 100) : "0%");
+            urlAssessmentResponse.setGoogleScore(googlePasses > 0 ? String.format("%.2f%%", (float) (googlePasses / googleCount) * 100) : "0%");
             urlAssessmentResponse.setGoogleTerms(termAssessments.stream()
                     .filter(term -> term.getSearchType() == TermAssessment.SearchType.GOOGLE)
                     .sorted(
